Buildable-stage
After PC5 merge, getting it to where it compiles properly.

RC had to include GMP/boost when building othrwise it would miss the library when touching other setups. Setup can be imporved to 'library' format and what not, noted.

RC edits are sketchy-ish.......and the particle update are REALLY bad.

Rework them!!!!

Also, git merge is quite buggy. Have to do a manual run-over every edit once the stuff launches and what not.